spring version down-grade and openFeign imple

+++ Spring_JobApplication/src/main/java/com/bound/test/dtos/JobDto.java
@@ -13,5 +13,5 @@ public class JobDto {
     private String description;
     private String location;
     private Double salary;
-    private Long companyId;
+    private Integer companyId;
 }
